Bug fix from moving the CURRENT_REQUEST_KEY
