+++ lib/torii/redirect-handler.js
@@ -7,7 +7,7 @@
  */
 
 import PopupIdSerializer from "./lib/popup-id-serializer";
-import { CURRENT_REQUEST_KEY } from "./services/popup";
+import { CURRENT_REQUEST_KEY } from "./mixins/ui-service-mixin";
 
 var RedirectHandler = Ember.Object.extend({
